@@ -28,6 +28,8 @@ import net.tridentsdk.factory.Factories;
 import net.tridentsdk.meta.MessageBuilder;
 import net.tridentsdk.meta.nbt.CompoundTag;
 import net.tridentsdk.server.TridentServer;
+import net.tridentsdk.server.data.MetadataType;
+import net.tridentsdk.server.data.ProtocolMetadata;
 import net.tridentsdk.server.entity.EntityBuilder;
 import net.tridentsdk.server.netty.ClientConnection;
 import net.tridentsdk.server.netty.packet.Packet;
@@ -45,8 +47,7 @@ import net.tridentsdk.world.LevelType;
 
 import javax.annotation.concurrent.ThreadSafe;
 import java.util.*;
-
-import static net.tridentsdk.server.data.ProtocolMetadata.MetadataType;
+import java.util.stream.Stream;
 
 @ThreadSafe
 public class TridentPlayer extends OfflinePlayer {
@@ -92,32 +93,29 @@ public class TridentPlayer extends OfflinePlayer {
 
         p.name = name;
 
-        p.executor.execute(new Runnable() {
-            @Override
-            public void run() {
-                p.connection.sendPacket(new PacketPlayOutJoinGame().set("entityId", p.entityId())
-                        .set("gamemode", GameMode.CREATIVE)
-                        .set("dimension", p.world().dimension())
-                        .set("difficulty", p.world().difficulty())
-                        .set("maxPlayers", (short) 10)
-                        .set("levelType", LevelType.DEFAULT));
+        p.executor.execute(() -> {
+            p.connection.sendPacket(new PacketPlayOutJoinGame().set("entityId", p.entityId())
+                    .set("gamemode", GameMode.CREATIVE)
+                    .set("dimension", p.world().dimension())
+                    .set("difficulty", p.world().difficulty())
+                    .set("maxPlayers", (short) 10)
+                    .set("levelType", LevelType.DEFAULT));
 
-                p.gameMode = GameMode.CREATIVE;
-                p.abilities.instantBreak = 1;
-                p.abilities.flySpeed = 1;
-                
-                // DEBUG =====
-                p.setLocation(new Position(p.world(),0,255,0));
-                p.spawnLocation = new Position(p.world(), 0,255,0);
-                // =====
-                
-                p.connection.sendPacket(PacketPlayOutPluginMessage.VANILLA_CHANNEL);
-                p.connection.sendPacket(new PacketPlayOutServerDifficulty().set("difficulty", p.world().difficulty()));
-                p.connection.sendPacket(new PacketPlayOutSpawnPosition().set("location", p.spawnLocation()));
-                p.connection.sendPacket(p.abilities.asPacket());
-                p.connection.sendPacket(new PacketPlayOutPlayerCompleteMove().set("location",
-                        p.spawnLocation().add(new Vector(0, 255, 0))).set("flags", (byte) 1));
-            }
+            p.gameMode = GameMode.CREATIVE;
+            p.abilities.instantBreak = 1;
+            p.abilities.flySpeed = 1;
+
+            // DEBUG =====
+            p.setLocation(new Position(p.world(),0,255,0));
+            p.spawnLocation = new Position(p.world(), 0,255,0);
+            // =====
+
+            p.connection.sendPacket(PacketPlayOutPluginMessage.VANILLA_CHANNEL);
+            p.connection.sendPacket(new PacketPlayOutServerDifficulty().set("difficulty", p.world().difficulty()));
+            p.connection.sendPacket(new PacketPlayOutSpawnPosition().set("location", p.spawnLocation()));
+            p.connection.sendPacket(p.abilities.asPacket());
+            p.connection.sendPacket(new PacketPlayOutPlayerCompleteMove().set("location",
+                    p.spawnLocation().add(new Vector(0, 255, 0))).set("flags", (byte) 1));
         });
 
         return p;
@@ -135,7 +133,9 @@ public class TridentPlayer extends OfflinePlayer {
     }
 
     @Override
-    protected void updateProtocolMeta() {
+    protected void encodeMetadata(ProtocolMetadata protocolMeta) {
+        super.encodeMetadata(protocolMeta);
+
         protocolMeta.setMeta(10, MetadataType.BYTE, skinFlags);
         protocolMeta.setMeta(16, MetadataType.BYTE, 0); // hide cape, might need changing
         protocolMeta.setMeta(17, MetadataType.FLOAT, 0F); // absorption hearts TODO
@@ -187,26 +187,16 @@ public class TridentPlayer extends OfflinePlayer {
         ticksExisted.incrementAndGet();
     }
 
-    public void setSkinFlags(final byte flags) {
-        this.executor.execute(new Runnable() {
-            @Override
-            public void run() {
-                skinFlags = flags;
-            }
-        });
+    public void setSkinFlags(byte flags) {
+        this.executor.execute(() -> skinFlags = flags);
     }
 
     /*
      * @NotJavaDoc
      * TODO: Create Message API and utilize it
      */
-    public void kickPlayer(final String reason) {
-        this.executor.execute(new Runnable() {
-            @Override
-            public void run() {
-                TridentPlayer.this.connection.sendPacket(new PacketPlayOutDisconnect().set("reason", reason));
-            }
-        });
+    public void kickPlayer(String reason) {
+        this.executor.execute(() -> connection.sendPacket(new PacketPlayOutDisconnect().set("reason", reason)));
     }
 
     public PlayerConnection connection() {
@@ -214,15 +204,13 @@ public class TridentPlayer extends OfflinePlayer {
     }
 
     public void setSlot(final short slot) {
-        this.executor.execute(new Runnable() {
-            @Override
-            public void run() {
-                if ((int) slot > 8 || (int) slot < 0) {
-                    TridentLogger.error(new IllegalArgumentException("Slot must be within the ranges of 0-8"));
-                }
-
-                TridentPlayer.super.selectedSlot = slot;
+        this.executor.execute(() -> {
+            if ((int) slot > 8 || (int) slot < 0) {
+                TridentLogger.error(new IllegalArgumentException("Slot must be within the ranges of 0-8"));
+                return;
             }
+
+            TridentPlayer.super.selectedSlot = slot;
         });
     }
 
@@ -231,10 +219,12 @@ public class TridentPlayer extends OfflinePlayer {
     }
 
     public void setFlying(boolean flying) {
-        this.flying = flying;
+        executor.execute(() -> {
+            this.flying = flying;
 
-        abilities.flying = (flying) ? (byte) 1 : (byte) 0;
-        connection.sendPacket(abilities.asPacket());
+            abilities.flying = (flying) ? (byte) 1 : (byte) 0;
+            connection.sendPacket(abilities.asPacket());
+        });
     }
 
     public boolean isFlying() {
@@ -263,18 +253,11 @@ public class TridentPlayer extends OfflinePlayer {
 
     @Override
     public void sendRaw(final String... messages) {
-        this.executor.execute(new Runnable() {
-            @Override
-            public void run() {
-                for (String message : messages) {
-                    if (message != null) {
-                        TridentPlayer.this.connection.sendPacket(
-                                new PacketPlayOutChat().set("jsonMessage", message)
-                                        .set("position", PacketPlayOutChat.ChatPosition.CHAT));
-                    }
-                }
-            }
-        });
+        executor.execute(() -> Stream.of(messages)
+                .filter((m) -> m != null)
+                .forEach((message) -> connection.sendPacket(new PacketPlayOutChat()
+                        .set("jsonMessage", message)
+                        .set("position", PacketPlayOutChat.ChatPosition.CHAT))));
     }
 
     public void sendChunks(int viewDistance) {
