@@ -25,23 +25,14 @@ import net.tridentsdk.entity.Projectile;
 import net.tridentsdk.entity.living.Mooshroom;
 import net.tridentsdk.entity.living.Player;
 import net.tridentsdk.event.entity.EntityDamageEvent;
+import net.tridentsdk.server.entity.TridentAgeable;
 import net.tridentsdk.server.entity.TridentLivingEntity;
 
 import java.util.UUID;
 
-public class TridentMooshroom extends TridentLivingEntity implements Mooshroom {
+public class TridentMooshroom extends TridentCow implements Mooshroom {
     public TridentMooshroom(UUID id, Position spawnLocation) {
         super(id, spawnLocation);
-    }
-
-    @Override
-    public int age() {
-        return 0;
-    }
-
-    @Override
-    public void setAge(int ticks) {
-
     }
 
     @Override
@@ -71,21 +62,6 @@ public class TridentMooshroom extends TridentLivingEntity implements Mooshroom {
 
     @Override
     public Player lastPlayerDamager() {
-        return null;
-    }
-
-    @Override
-    public boolean isNameVisible() {
-        return false;
-    }
-
-    @Override
-    public void applyProperties(EntityProperties properties) {
-
-    }
-
-    @Override
-    public <T extends Projectile> T launchProjectile(EntityProperties properties) {
         return null;
     }
 
